fix(ast): correctly mark fields as nullable when not required
Creates a new type alias `Nullable<T>` which will make a given field nullable
if that field is not required by the schema.

This more accurately models the allowed types for GraphQL since GraphQL will
accept `null` as a valid value when a field is marked as such. Prior to this change,
`null` was not allowed as a value in TypeScript since the generated type definition did not allow it.

Closes #1128