feat(plugin): Update plugin with null union support
Prior to these changes, the GraphQL plugin only supported decorating field properties when the
type of a class property was optional. However, sometimes typescript users like to use union
types that include null instead to indicate a nullable field. This change makes it possible to
support nullable fields when the union includes a null type.