fix(plugin) downlevel implicit registerEnumType import in CommonJS
When code is downleveled to ES5 it automatically change all
usages of imported members except those which was added by transform.

This breaks `registerEnumType()` calls added by transform.

The solution is create an explicit dedicated import only for this method in the
transform and not relay on exisitng imports at all.