fix(@nestjs/graphql): set unique ids on resolvers to fix cache errors
The resolver metadata cache currently uses the resolver name for the
key but this causes issues if we have two resolvers in different modules
but with the same name and identical method signatures. For controllers,
each controller gets a unique ID that is used as the key in the cache.
Add code to set up unique IDs for resolvers as well so that cache
retrieval is correct.

Fixes nestjs/graphql#1907