fix: metadata storage refactor caused a bug in field resolver
When creating a field resolver with an input we reached a bad edge case where the fields array was fetched via reference instead of an immutable array. This caused artifacts appearing in the array and bad metadata appeared when trying to parse the schema.

Returning a new array from the metadata instead of the reference for the existing array solves this issue